Lib

Description

This module provides convenience functions built from the core of the RASP-L language.

It is based on Listing 3 of "What Algorithms Can Transformers Learn", https://arxiv.org/abs/2310.16028, by Zhou et al.

Logical Operations

(?) :: [Bool] -> (Sequence, Sequence) -> Sequence

Use a boolean sequence to select between two sequences. Also known in Python RASP-L as "where", which is a restricted keyword in Haskell.

where' :: [Bool] -> Sequence -> Sequence -> Sequence

Use a boolean sequence to select between two sequences. Provided for compatibility with Listing 3, but with an apostrophe to avoid a name clash with the "where" keyword.

shiftRight :: Token -> Int8 -> Sequence -> Sequence

Shift a sequence to the right by a given number of elements, filling the vacated positions with the provided token.

toBool :: Token -> Bool

mask :: Token -> [Bool] -> Sequence -> Sequence

Running Aggregation

cumSum :: [Bool] -> Sequence

maximum' :: Sequence -> Sequence

minimum' :: Sequence -> Sequence

argmax :: Sequence -> Sequence

argmin :: Sequence -> Sequence

Token Comparison

leq :: Token -> Token -> Bool

geq :: Token -> Token -> Bool

lt :: Token -> Token -> Bool

gt :: Token -> Token -> Bool

Sampling

sample

Arguments

:: Token
End of sequence token

-> (Sequence -> Sequence)
RASP-L program to extend the sequence

-> Sequence
Initial sequence

-> Word8
Number of steps to decode

-> Sequence

Greedily and autoregressively sample the output of a RASP-L program on a sequence.